Add generic exception mapper for REST API
