+++ src/org/opensolaris/opengrok/web/api/constraints/PositiveDuration.java
@@ -20,7 +20,7 @@
 /*
  * Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
  */
-package org.opensolaris.opengrok.web.constraints;
+package org.opensolaris.opengrok.web.api.constraints;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
