+++ src/org/opensolaris/opengrok/web/api/error/ValidationExceptionMapper.java
@@ -20,24 +20,23 @@
 /*
  * Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
  */
-package org.opensolaris.opengrok.web.constraints;
+package org.opensolaris.opengrok.web.api.error;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
-import javax.validation.constraintvalidation.SupportedValidationTarget;
-import javax.validation.constraintvalidation.ValidationTarget;
-import java.time.Duration;
+import javax.validation.ValidationException;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.ext.ExceptionMapper;
+import javax.ws.rs.ext.Provider;
 
-@SupportedValidationTarget(ValidationTarget.ANNOTATED_ELEMENT)
-public class PositiveDurationValidator implements ConstraintValidator<PositiveDuration, Duration> {
-
-    @Override
-    public void initialize(final PositiveDuration positiveDuration) {
-
-    }
+@Provider
+public class ValidationExceptionMapper implements ExceptionMapper<ValidationException> {
 
     @Override
-    public boolean isValid(final Duration duration, final ConstraintValidatorContext context) {
-        return duration != null && !duration.isNegative() && !duration.isZero();
+    public Response toResponse(final ValidationException e) {
+        return Response.status(Response.Status.BAD_REQUEST)
+                .entity(e.getMessage())
+                .type(MediaType.TEXT_PLAIN)
+                .build();
     }
+
 }
